Update dataflow to come from Main component
Split data to reside in each charts initally because
was having difficulty getting setState to update
correctly within nested data structures, which a required
format for the C3 component.

Changed the column data to be stored in just arrays in
a flat manner and created required structure in a new
method. Added benefits will be allowing one AJAX
call and only needing one ChartInstance component
instead of separate for Daily and Cumulative charts.

Also, will be good when syncing with submit button which
will need to take a callback for updating the data
of the parent component.